refactor cards into subfolder with seperated .styles file if needed
