+++ src/components/cards/cardPost/cardPost.styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { TextMuted } from '../../ui/content/textMuted';
+
+export const SubHeader = styled(TextMuted)`
+  font-size: var(--font-size-smaller);
+`;
+
+SubHeader.displayName = 'SubHeader';
