+++ src/components/cards/cardProject/cardProject.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import { Card } from '../ui/container/card';
-import { Column } from '../ui/container/column';
-import { Fill } from '../ui/container/fill';
-import { Row } from '../ui/container/row';
-import { Tag } from '../ui/content/tag';
-import { MarginSmall } from '../ui/margins/marginSmall';
-import { UnstyledLink } from '../ui/content/unstyledLink';
-import { animationSpring } from './animations';
-import { CardHeader } from './shared.styles';
+import { Card } from '../../ui/container/card';
+import { Column } from '../../ui/container/column';
+import { Fill } from '../../ui/container/fill';
+import { Row } from '../../ui/container/row';
+import { Tag } from '../../ui/content/tag';
+import { MarginSmall } from '../../ui/margins/marginSmall';
+import { UnstyledLink } from '../../ui/content/unstyledLink';
+import { animationSpring } from '../animations';
+import { CardHeader } from '../shared.styles';
 
 interface CardProjectProps {
   title: string;
